+++ bin/constants.py
@@ -1,0 +1,13 @@
+# Default parameters
+FLANK = 500
+FUSION = False
+TARGET_LEAD = 50
+NR_PRIMERS = 4
+ALLOWED_MISMATCHES = 5
+MAX_MAPPINGS = 5
+COLOURS = [[255, 0, 0],  # red
+           [0, 255, 0],  # green
+           [0, 0, 255],  # blue
+           [255, 0, 255],  # Pink
+           [0, 255, 255],  # cyan
+           [255, 255, 0]]  # yellow
